backend/requirements.txt
# Google ADK and AI dependencies
google-genai>=0.2.0
google-adk>=0.1.0

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# MCP (Model Context Protocol)
mcp>=0.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
